Fix format of change-log
Each version-numbers is now links to its tag on GitHub, rather then
being followed by a visible URL to that tag.

Brings in line with the change-log format of other Stripes projects. I
can only assume that this project's divergence from that format was
inadvertent in the first release, and has been unthinkingly followed
(incuding by me) in subsequent releases.