show-single-result should be optional.
The show-single-result behavior should be optional and default to false.
This restores the previous behavior by default and gives apps the
option of using it on request.

It causes problems in the ui-plugin-find-user context where winnowing a
search in a modal window causes the parent window to redirect to the
entry's detail record when, instead, we simply want to indicate that the
search returned a single record and the user should click it to select
it.

Fixes STSMACOM-52. Refs UIIN-58, UIREQ-60, UICHKOUT-54, UIU-373.