Guard against selecting row when no mutations available
I don't understand why the `successfulMutations` array would be empty
after creating a user, and indeed I never see this condition on my
development box -- which makes it very hard to debug -- but it does
happen on folio-testing and snapshot. So I've added a guard, whose
effect I can't observe on my own box, in the hope that it will make
the public boxes work right when it gets uploaded.

Yeah, I know.

Should fix STSMACOM-83.