Always display sole hit when narrowing down to one result
This wasn't working consistently when using GraphQL, because Apollo
briefly sets result-count to `null` while it's in the process of
running a search, so we weren't seeing the expected transition between
a hit-count of more than one and a hit-count of exactly one. (It was
still working _sometimes_ because if the single-hit search is in the
cache from earlier, Apollo leaps straight to it without going through
an in-progress stage.)

This was fixed by remembering the last _non-null_ result-count, and
showing the full record when _that_ goes from greater than one to
exactly one.

Fixes STSMACOM-98.