optionally hide the editable list
The `listSuppresssor` function and `listSuppressorText` props may be
used to completely suppress the dispaly of the EditableList component.
This is useful when, for example, the component is used in conjunction
with `props.rowFilter` and a valid selection must be made from the
rowFilter before any element can be added/updated/removed from the list.

Refs UIORG-82, UIORG-83