cache user metadata
Cache user metadata so we don't look up the same values over and over.
Instead of looking up the same values over and over, cache the user
metadata. In order to avoid doing the same lookups of user metadata
multiple times, the results is stored in a cache. By storing a map of
user-id to user-object, we can avoid repeated lookups.

Refs STCOM-308