wait for settings query before rendering form (#237)
Previously, the form would render immediately upon invocation and then
rerender when the settings resource returned. If you are fast with the
mouse, and I mean real fast like faster than The Waco Kid fast, or
maybe if you are a Nightmare test, you could click the submit button
between the first (pre-resource results) and second (post-resource
results) renders. Without an existing record with an id to coalesce into
(line 59) we'll get a POST request instead of a PUT (line 65) that will
blow up when Okapi rejects the duplicate record.

This solves all that.

Refs UICIRC-75