make stripes-core a peerDependency
stripes-core must be a peerDependency so it is included once per bundle.
Listing it as a regular dependency means it may be nested, rather than
hoisted, which can lead to multiple versions, but separate versions will
have separate instances of context, resulting in context-consumers that
can't see the context offered by the context-providers they are trying
to subscribe too.

Refs STRIPES-557