don't sort ControlledVocabulary queries (#317)
It used to be the case that if you tried to run a query with an index
that didn't exist the parser would realize this and just ignore the
index. Now, it throws a query validation exception. Probably, this is a
good thing; it's good to be strict about what we're running. But at this
late stage, when things were working forever, it feels like you're just
about to submit your problem set and along comes the TA who says, "You
know you have to compile with '-Wall', right?" <sigh>

Refs MODUSERS-98