indicate whether `<EntrySelector>` has any entries
Use a CSS class to indicate when this component is rendered with one or
more entries. This component may re-render once results become
available, but there's no easy way for a test to know that because there
are no element classes, no element ids, nothin'. It would have to parse
some crazy long nested structure of divs and anchors, which feels pretty
nasty.

In fact, the loan renewal tests in platform core sometimes blow up when
trying to click on non-existant links because they don't have a good way
to test whether those links are present. This helps.

Refs UITEST-55