validate callouts before calling them
Sometimes, Nightmare tests would blow up with the following error:
```
EntryWrapper.js:145 Uncaught (in promise) i
TypeError: Cannot read property 'sendCallout' of null
```
which is the result of the delete method being called before the whole
DOM renders, which causes the callout to fail because it relies on a ref
which isn't available yet because the DOM hasn't rendered. Got that?
Good.