improve metadata display (#479)
Complete reworking of the metadata display. The previous implementation
was so broken it's not even worth trying to describe how it's changed
except to say that it doesn't suck anymore.

This implementation gathers the list of updaters from the list of
entries and then makes a single query to the `users` endpoint to
retrieve all the records. It also checks permissions _before_ doing so,
in order to avoid attempting (and failing) to retrieve data the end-user
doesn't have permission to read.

Refs STSMACOM-181, STSMACOM-182, STSMACOM-183, UIU-873, UIIN-462.