Remove .isRequired from 'records' propType
The actual code already allows for the situation where no value is
passed in for `records`, i.e. the top level of the JSON returned from
the WSAPI is an array of records rather than an object containing a
member whose value is this array. Since this works correctly, a valid
invocation should not yield a console error.