Pass clientGeneratePk prop into manifest option
When this is turned on, a caller can pass in a falsy value such as an
empty string to modify the behaviour of stripes-connect, allowing the
back-end module to provide the ID of newly created records. This
allows `<ControlledVocab>` to work correctly with back-end modules
such as mod-rs and mod-directory created using Knowledge Integration's
toolkit.

The `clientGeneratePk` prop explicitly defaults to true, matching the
previous implicit default.

Requires stripes-connect ^5.1.2.