*sigh* fix propTypes declaration for resources.updaterId (#553)
This is a double fix. First, I incorrectly specified it as type string
rather than arrayOf(string). But also, it seems that early in the
lifecycle it gets furnished to this component as an empty _object_
rather than array -- I suspect because this is stripes-connect's
default. So we use oneOfType to accept either.