STSMACOM-230 ViewMetaData shouldn't fetch constantly (#579)
Under some circumstances, `<ViewMetaData>` would get stuck in a loop and
React would eventually throw a Maximum Update Depth Exceeded error.
Debugging showed that it was receiving new props for creator-id and
updater-id but that, somehow, only the creator-id was being copied into
the `updaterIds` local resource. This meant the comparison of the props
to the resource always returned false and the query ran on each
iteration. Yuck.

Fixes STSMACOM-230