provide yarn.lock to assist with tests
Unit tests consistently succeed locally but fail in CI. It has been our
standard to _avoid_ comitting `yarn.lock` files to modules in order to
avoid confusion about how they work with regard to dependency resolution
in the context of a platform (i.e. they will be ignored in that context,
hence their only value is to ensure consistency during testing). Here,
however, since testing is what is failing, it seems like providing the
file may be worth a shot.