Helper-app rendering no-ops if the helper is undefined
The previous version contained what looks like a straight-up bug where
the rendering no-opped only if the helper _name_ was undefined. This
meant you could invoke a URL with, say, `helper=noSuchThing`, and
`<SearchAndSort>` would fail to find that helper but still try to
render the component that implements it, even though it was undefined.

Thanks to @EthanFreestone for helping me spot this.