<Settings> supports new optional `additionalRoutes` property
If provided, this is an array of `<Route>` objects that are included
in the `<Settings>` component's routing for its sub-pane, but which
are not displayed in the visible list of settings sections.

This is useful for routing paths below the top level of a settings
page. For example, if you want your Foo module's Bar settings page to
be able to see a specific Bar in detail, you might pass `pages`
including:

	{
	  route: 'bar',
	  id: 'Maintain all Bars',
	  component: ListOfBar,
	},

and also `additionalRoutes` including:

      <Route
        path={`${match.path}/bar/:id`}
        render={p => <SingleBar {...p} label="Maintain specific Bar" />}
      />