UIU-1641: New/Edit User Record: Primary address toggle uses invalid ARIA
After looking up how the aria-labelledby attribute is supposed to be used,
I decided to remove this attribute entirely from the radio button. This
attribute is supposed to tell assistive technologies what label is associated
with this form control.  But the label in question already wraps around the
control, making it clear from context what form control it's associated with.
The label/control shouldn't need any attribute to associate the two.

Refs: https://issues.folio.org/browse/UIU-1641