Add ClipCopy component (#858)
Display a clickable-clipboard to copy text to the clipboard.

I tried hard to get tests working for this component, honest, but 
never met with consistent success. There are some notes on 
`tests/ClipCopy-test.js` but the short version is that things just 
weren't reliable. Talking with @JohnC-80, we wonder if the 
problem is context not being available immediately after mount, 
but who knows? Not me, that's for sure.

Refs UIREQ-478

Co-authored-by: Viktor Soroka <viktor_soroka@epam.com>