STSMACOM-401: failed tag-creation shows a bogus "Tag created" success toast
I addressed this by using then/catch/finally syntax to provide error
handling for the promise when the attempt to save tag data fails.  It now
sets a variable that the .finally clause can check and use to decide
wether or not to show the success toast.  I also turned off the ugly
error that used to pop up if you could not save tag data. The user
should now see error toasts that indicate if the save attempt failed,
and I added new labels to the translation files to support this.

Refs: https://issues.folio.org/browse/STSMACOM-401