STSMACOM-463 refactor ClipCopy with useIntl
There was a bad interaction between `<FormattedMessage>` and
`useCallback` in `<ClipCopy>`, likely because `<FormattedMessage>`
implicitly relies on the `<IntlProvider>` context being available, but
this is not passed into `useCallback`.

Refactoring with `useIntl`, allowing the `intl` object to become an
explicit dependency of the `useCallback` hook, resolves this issue.

We moved the `CalloutContext` up a level in the component hierarchy in
STCOR-481, which is probably the root cause of this.

Refs STSMACOM-463, UIIN-1357, UIOR-635, STCOR-481