STSMACOM-490 export missing ConfigManager components
`ConfigForm` was replaced with `ConfigReduxForm` (and a parallel
component `ConfigFinalForm`) but (a) neither was exported and (b)
`ConfigForm` continued to be exported but now is an entirely different
component.

Rather than try to restore things as they were since use of `ConfigForm`
appears to be confined to a couple of pages in ui-developer, this PR
simply exports `ConfigReduxForm` and `ConfigFinalForm`, restoring the
old behavior but under a new name. That's good enough, and small enough,
and gosh darn it, people should like it.

Refs STSMACOM-490, UID-72