lint Notes
Since we get HTML from react-quill in notes, we need to either display
the editor in read-only mode, which requires theming, or use
`dangerouslySetInnerHtml` as here. I wish there were a simpler way to
display a read-only pane from react-quill, but it isn't clear that there
is, and using `danger...` seems to be the agreed-upon [easiest solution](https://stackoverflow.com/questions/40952434/how-do-i-display-the-content-of-react-quill-without-the-html-markup).