Linty mclintface (#1158)
* lint CustomFields
Refs should not be provided in the dependency array for `useCallback`
since the entire point of refs is that they _don't_ trigger a re-render
the way a state-change will, and the entire point of `useCallback`'s
dependency array is to collect the list of changes that, during a
render, should cause this function to run. As lint says:
  > warning  React Hook useEffect has an unnecessary dependency: 'paneTitleRef.current'. Either exclude it or remove the dependency array. Mutable values like 'paneTitleRef.current' aren't valid dependencies because mutating them doesn't re-render the component  react-hooks/exhaustive-deps

* lint LocationLookup
props sort is pretty self explanatory

* lint Notes
Since we get HTML from react-quill in notes, we need to either display
the editor in read-only mode, which requires theming, or use
`dangerouslySetInnerHtml` as here. I wish there were a simpler way to
display a read-only pane from react-quill, but it isn't clear that there
is, and using `danger...` seems to be the agreed-upon [easiest solution](https://stackoverflow.com/questions/40952434/how-do-i-display-the-content-of-react-quill-without-the-html-markup).

* lint SearchAndSort
provide missing effect dependency
trim whitespace