STSMACOM-639 correctly provide default sortby value
As noted by helpful comments from @miketaylor in this file, react-redux
does not receive defaultProps values, and hence can't pass them onto
stripes-connect for substitution in a manifest. T

The situation we ended up in was that our default value was being
ignored, causing evaluation of the manifest to fail and thus the query
would never fire.

Refs STSMACOM-639