More configurable makeQueryFunction (#1350)
The boolean `escape` parameter is replaced by a `config` object, which can contain multiple configuration options. These at present include:
* `escape` (boolean) — as before, specifies whether to escape quote and backslash values in the query [default: `true`]
* `rightTrunc` (boolean) — whether to append `*` to query terms [default: `true`].

And of course it will be easy to add other such entries later if we need them.

This version of the function is 100% backwards-compatible with the previous version: a boolean can be passed in place of the config object, and is interpreted as the value of `config.escape`. As before, this defaults to true if not specified.