restore original signature of makeOkapiRequest
Careful inspection of the console logs showed a miragejs error coming
from the a CSS file (what the???) that complained about `namespace`
being unset. I don't understand why this is coming from a CSS file, nor
why changing the dependency array to `useCallback` prompted this error,
but the mirage docs at
https://miragejs.com/api/classes/server/#namespace demonstrate how it
can be used, and it works here. Weird.

I wish I understood this better. My hunch is that something deep in
mirage expects `namespace` to be initialized, and somehow it's getting
unset in our code, or we have some ancient BTOG dep that is now out of
sync with the current miragejs release.