Optional `resourcePath` prop for `<EntryManager>` (#1361)
The `<EntryList>` component was created on the assumption that the records it manages are returned in full when a list is requested: so, for example, the list of records returned from `/path/to/endpoint` are each the same shape as the individual ones obtained by ID at paths like `/path/to/endpoint/12335`. This pattern is generally followed by most web-service endpoints provided by most FOLIO modules.

Some modules, however, return a list of summary records (for example, just ID and title), so that it's necessary to fetch each full record separately by ID. When `<EntryList>` is used to manage records on such an endpoint, the path to that service should be passed in as the `resourcePath` prop. When this is done, the full record is fetched each time a user views or edits a record.

Fixes STSMACOM-757.

This is required by:
* https://issues.folio.org/browse/UIHAADM-7
* https://issues.folio.org/browse/UIHAADM-8
* https://issues.folio.org/browse/UIHAADM-9