`<EntryManager>`: `parseInitialValues` prop works with `resourcePath`
The `parseInitialValues` function is now applied correctly when
`resourcePath` is defined (see STSMACOM-757) so that full records are
loaded from a WSAPI.

Previously, the transformation would be applied to the brief version
of the record before the full record was loaded, so that fields that
occur only in the full version of the record were not transformed.

Also, I have documented these previously undocumented props.

Fixes STSMACOM-761.