test: Test fixes
I think I found a long standing bug in the test, in ERM we have long needed to set initialSearchState to  { query: '', qindex: ''} in order for resetAll to take effect. This is not reflected in the test, causing it to fail on reset.   I believe the test previously passed via some fluke, since the Button in the test handler was not using onClick and the search submit is what has actually changed.

This ought to be double checked though