STSMACOM-775: Search And Sort Query - fire queryStateReducer events on search submit (#1388)
* feat: Add SASQ event for search submit and all submit

Added internalSetState call for submit all and submit search, which does not change internal state itself, but fires an event that queryStateReducer can use to trigger changes from those events

* chore: Callbacks

Ensure that callQuerySetter is called as callback to the internal setState, so it has the benefit of any changes made within queryStateReducer

* test: Test fixes

I think I found a long standing bug in the test, in ERM we have long needed to set initialSearchState to  { query: '', qindex: ''} in order for resetAll to take effect. This is not reflected in the test, causing it to fail on reset.   I believe the test previously passed via some fluke, since the Button in the test handler was not using onClick and the search submit is what has actually changed.

This ought to be double checked though

* chore: DebouncedCallQuerySetter

Swapped onSubmitSearch back to debouncedCallQuerySetter for consistency with the old code

STSMACOM-775