STSMACOM-785 add default parameters to onSubmit, submitAll internal handlers (#1404)
d197a0 in PR #1388 added some necessary event handling and changeTypes
to SASQ internals, but inadvertently changed default values that
external code relied on. 

State is initialized with `changeType = init.reset`, but when SASQ
has `setQueryOnMount = true` or an `initialSearchState` applied, it
calls `submitAll()` on mount. Previous to d197a0, `submitAll()` had
no `changeType` so the initial `changeType` persisted. After d197a0,
the `changeType` was set to `all.submit`, over-writing `init.reset`.

This code adds the `changeType` as a default parameter for the methods
that we would expect, with overridden `changeType`s applied for those
situational cases where `submitAll()` and `submitSearch` are called - that
is, at mount time, given the previously mentioned props, and upon a
user-triggered reset.

Refs STSMACOM-785, STSMACOM-775