STSMACOM-822: Do not trigger logic for auto-opening the record's view screen in `<SearchAndSort>` if URL already contains the record ID. (#1459)
Description
When there is a single search result and the third pane is open, and
then the user creates a new record in a different route and when the new
record is created the user is redirected to the main page (with the
previously opened third pane). The third pane doesn't contain the newly
created record, it contains the previously opened one.

Root cause:
The logic for automatically opening the record’s view screen fires even
if the URL already contains the record’s view screen route. The logic
must only be triggered if there is no open record.

Approach
Add a check for a record ID in the URL. If it exists, then prevent the
automatic opening of the record.

(cherry picked from commit fbc8693a902be67e57780ae0a80de1fd02c38e36)