STSMACOM-836 omit "+ New" button when user lacks permission (#1477)
Omit the "+ New" button, rather than disabling it, when the user lacks
permission. The convention is to omit unusable elements rather than to
disable them. An element is present-but-disabled only if taking some
other action available on the same page will enable it (e.g. this button
becomes disabled when another element is being edited or added, and then
becomes enabled again on save).

Refs STSMACOM-836