STSMACOM-893 EntryManager sort by id if sort-key is absent (#1561)
Avoid an NPE when sorting lists if the requested sort-key is missing by
using `id` as the key instead. Should we check for `id`? Eh, maybe, but
`<EntryManager>` is a tool for handling entries retrieved from API
requests so it seems like a safe assumption.

This bug was introduced in 1a8f5450b566c0cf21626760e2ef40de5e63f8c2.

Refs STSMACOM-893