STSMACOM-903 Pass props for rendering aside content for custom fields accordion (#1578)
https://folio-org.atlassian.net/browse/STSMACOM-903

## Purpose

Since the `<EditCustomFieldsRecord>` component includes an
`<Accordion>`, it should support passing content as props based on the
accordion’s state — `displayWhenClosed` and `displayWhenOpen`.
The prop names are consistent with those used in the `<Accordion>`
component.

As per the current business requirement, this accordion might be marked
as _hidden_ within the order template. This designation allows the
component to be excluded from the rendering of both the PO and POL form
and view. (Ref https://folio-org.atlassian.net/browse/UIOR-1394)

## Screenshots



https://github.com/user-attachments/assets/d4986439-8008-4701-bf18-9713213e2021