STSMACOM-790: Add `configNamePrefix` prop to custom fields components to be able to store section titles separately for different `entityType`s. (#1584)
## Description
When multiple instances of the `EditCustomFieldsSettings`/
`EditCustomFieldsRecord`/ `ViewCustomFieldsRecord`/
`ViewCustomFieldsSettings` UI component, each associated with a distinct
`entityType`, are used within a single application, then modifying the
accordion label in one component results in an unintended change to the
label in all other components. This behavior contradicts the expected
functionality, causing inconsistency in the labeling of accordion
sections based on entityTypes.

The issue arises due to the request sent to the configurations API to
store the accordion label, which utilizes a single `configName` for all
entityTypes.

Using a single `configName` for multiple entityTypes, leads to a
universal change in the accordion label across all instances of the
`EditCustomFieldsSettings`/ `EditCustomFieldsRecord`/
`ViewCustomFieldsRecord`/ `ViewCustomFieldsSettings` components,
regardless of their associated entityTypes.

## Approach
Use an optional `configNamePrefix` property to extend `configName`.

## Issues
[STSMACOM-790](https://folio-org.atlassian.net/browse/STSMACOM-790)