STSMACOM-913 added a new getCustomErrorMessages prop to EditableListForm (follow-up) (#1590)
[Previous
PR](https://github.com/folio-org/stripes-smart-components/pull/1589)
involved providing a custom error handler to `<ControlledVocab>`, but it
caused list of items to not refresh when an item creation failed.
Instead the new approach is to provide an error message getter via
props. This way when an item creation fails - the list is still
refreshed.

## Screenshots


https://github.com/user-attachments/assets/5c666a3a-9a0b-4313-9a6e-70e230bbda85

## Issues
[STSMACOM-913](https://folio-org.atlassian.net/browse/STSMACOM-913)