STSMACOM-914: Settings Custom Fields View - include the `Display in accordion`, `Hidden` fields, and re-order the fields. (#1591)
## Description
Settings Custom Fields View:
- added new property `hasDisplayInAccordionField` to show/hide the new
`Display in accordion` field, which is designed to store the name/id of
the accordion where the custom field will be placed. Currently, BE is
done only for the mod-users, mod-orders, or any other module should do
some extra work to use this field;
- help text field is now always displayed instead of when value is
present;
- changed fields order to `Field label`, `Help text`, `Display in
accordion`, `Hidden`, `Required`;
- checkbox doesn't have `Required` field;
- created new `DisplayInAccordion` component for `Display in accordion`
field;
- created new `HiddenValue` component;
- added new `displayInAccordionOptions` property, it is a list of
options `[{ value, label }]`, `value` is used to find the correct
`label` and display it for the `Display in accordion` field;
- the initial `Custom fields` option label in the `Display in accordion`
field will be the same as section title.

## Issues
[STSMACOM-914](https://folio-org.atlassian.net/browse/STSMACOM-914)

## Screenshots


https://github.com/user-attachments/assets/ab9fb6b7-29d0-4d57-abc0-532b85fe729e


![image](https://github.com/user-attachments/assets/6fb25b59-89a1-407e-bf7d-0750784c2819)