STSMACOM-916: Settings Custom Fields Edit - include the "Display in accordion" field and re-order the fields. (#1593)
## Purpose
In custom fields settings, add the ability to select the name of the
accordion to which the custom field will be added. Currently, all custom
fields are added to a single accordion with the default name `Custom
fields`. BE only supports the Users app, so the new properties will be
optional. For other apps, the order of the fields has been changed.

`EditCustomFieldsSettings` component:
- added the `Display in accordion` field and the ability to edit it;
- reordered fields;
- added a new `hasDisplayInAccordionField` property to display/save the
`Display in accordion` field;
- added a new `displayInAccordionOptions` property for `dataOptions`;
- if the option matches the default one, the option label will match the
custom field accordion name.

## Issues
[STSMACOM-916](https://folio-org.atlassian.net/browse/STSMACOM-916)

## Screencasts




https://github.com/user-attachments/assets/966d51ed-9f02-4b53-84cb-a467df657696