STSMACOM-917: Validate custom fields by filtering out those with empty `refId` in `useCustomFieldsQuery`. (#1604)
## Description
The `EditCustomFieldsRecord` component crashes the entire page when the
BE returns custom fields with an empty `refId` value (`"refId": ""`).
This occurs because Final Form attempts to register fields with invalid
names.

When creating the _first_ custom field with a name of non-Latin
characters, BE returns an empty `refId`. BE has a [bug
](https://folio-org.atlassian.net/browse/FCFIELDS-70)related to this,
but it has been decided to also add protection against this issue in the
UI.

## Approach
Filter out invalid custom fields before they reach the UI components.

## Issues
[STSMACOM-917](https://folio-org.atlassian.net/browse/STSMACOM-917)